requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
Pillow==10.1.0
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2